Refactor: PR의Review반영 (#94)
match-gameMode.enum.ts -> match-game-mode.enum.ts로 변경

feat(#94)
